Move services to sub directories
So that test files can be paired with them nicely.
